# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * IGNOU PG Political Science Assignment Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Assignment {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  subject: string;
  year: string;
  fileName: string;
  createdAt: string;
}

export interface AssignmentInput {
  title: string;
  description?: string;
  subject: string;
  year: string;
  fileName: string;
  /** Base64 encoded file content */
  fileData: string;
}

export type PaymentRequestStatus = typeof PaymentRequestStatus[keyof typeof PaymentRequestStatus];


export const PaymentRequestStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface PaymentRequest {
  id: number;
  name: string;
  phone: string;
  utrNumber: string;
  status: PaymentRequestStatus;
  createdAt: string;
  updatedAt: string;
}

export interface PaymentRequestInput {
  name: string;
  phone: string;
  utrNumber: string;
}

export type PaymentStatusStatus = typeof PaymentStatusStatus[keyof typeof PaymentStatusStatus];


export const PaymentStatusStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface PaymentStatus {
  status: PaymentStatusStatus;
  name: string;
  phone: string;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminToken {
  token: string;
  message: string;
}

export interface AdminStats {
  totalPayments: number;
  pendingPayments: number;
  approvedPayments: number;
  rejectedPayments: number;
  totalAssignments: number;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export type CheckPaymentStatusParams = {
phone: string;
};

export type ListPaymentRequestsParams = {
status?: ListPaymentRequestsStatus;
};

export type ListPaymentRequestsStatus = typeof ListPaymentRequestsStatus[keyof typeof ListPaymentRequestsStatus];


export const ListPaymentRequestsStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;