# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: IGNOU PG Political Science Assignment Portal API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: assignments
    description: Assignment operations
  - name: payments
    description: Payment request operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List all assignments (public metadata)
      responses:
        "200":
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"

  /payment-requests:
    post:
      operationId: submitPaymentRequest
      tags: [payments]
      summary: Submit a payment request after UPI payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequestInput"
      responses:
        "201":
          description: Payment request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequest"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /payment-requests/check:
    get:
      operationId: checkPaymentStatus
      tags: [payments]
      summary: Check payment status by phone number
      parameters:
        - name: phone
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentStatus"
        "404":
          description: No payment request found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminToken"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/payment-requests:
    get:
      operationId: listPaymentRequests
      tags: [admin]
      summary: List all payment requests (admin only)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: List of payment requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentRequest"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/payment-requests/{id}/approve:
    patch:
      operationId: approvePaymentRequest
      tags: [admin]
      summary: Approve a payment request (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment request approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequest"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/payment-requests/{id}/reject:
    patch:
      operationId: rejectPaymentRequest
      tags: [admin]
      summary: Reject a payment request (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequest"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/assignments:
    post:
      operationId: createAssignment
      tags: [admin]
      summary: Create assignment metadata (admin only, file uploaded separately)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentInput"
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/assignments/{id}:
    delete:
      operationId: deleteAssignment
      tags: [admin]
      summary: Delete an assignment (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignment deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Assignment:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        subject:
          type: string
        year:
          type: string
        fileName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - subject
        - year
        - fileName
        - createdAt

    AssignmentInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        subject:
          type: string
        year:
          type: string
        fileName:
          type: string
        fileData:
          type: string
          description: Base64 encoded file content
      required:
        - title
        - subject
        - year
        - fileName
        - fileData

    PaymentRequest:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        utrNumber:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - phone
        - utrNumber
        - status
        - createdAt
        - updatedAt

    PaymentRequestInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        utrNumber:
          type: string
      required:
        - name
        - phone
        - utrNumber

    PaymentStatus:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected]
        name:
          type: string
        phone:
          type: string
      required:
        - status
        - name
        - phone

    AdminLoginInput:
      type: object
      properties:
        password:
          type: string
      required:
        - password

    AdminToken:
      type: object
      properties:
        token:
          type: string
        message:
          type: string
      required:
        - token
        - message

    AdminStats:
      type: object
      properties:
        totalPayments:
          type: integer
        pendingPayments:
          type: integer
        approvedPayments:
          type: integer
        rejectedPayments:
          type: integer
        totalAssignments:
          type: integer
      required:
        - totalPayments
        - pendingPayments
        - approvedPayments
        - rejectedPayments
        - totalAssignments

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message