# lib/api-zod/src/generated/types/adminToken.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * IGNOU PG Political Science Assignment Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminToken {
  token: string;
  message: string;
}