# lib/api-zod/src/generated/types/assignmentInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * IGNOU PG Political Science Assignment Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AssignmentInput {
  title: string;
  description?: string;
  subject: string;
  year: string;
  fileName: string;
  /** Base64 encoded file content */
  fileData: string;
}