# lib/api-zod/src/generated/types/paymentRequest.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * IGNOU PG Political Science Assignment Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentRequestStatus } from './paymentRequestStatus';

export interface PaymentRequest {
  id: number;
  name: string;
  phone: string;
  utrNumber: string;
  status: PaymentRequestStatus;
  createdAt: string;
  updatedAt: string;
}